fixing nav migration

+++ db/migrate/20100208212049_migrate_pages_to_navigation.rb
@@ -1,9 +1,13 @@
 class MigratePagesToNavigation < ActiveRecord::Migration
   def self.up
-    add_column :pages, :menus_count, :integer, :default => 0
+    unless ActiveRecord::Base.connection.tables.include?("menus")
+      add_column :pages, :menus_count, :integer, :default => 0
+    end
   end
 
   def self.down
-    remove_column :pages, :menus_count
+    unless ActiveRecord::Base.connection.tables.include?("menus")
+      remove_column :pages, :menus_count
+    end
   end
 end
